7일차 : 씬 그리드 추가

+++ EditorTool/EditorToolManager.h
@@ -11,19 +11,19 @@ public:
 	void Init();
 	void Update();
 
-	void SetSelectedObjH(int32 id) { _selectedIdH = id; }
-	void SetSelectedObjP(int32 id) { _selectedIdP = id; }
+	void SetSelectedObjH(int64 id) { _selectedIdH = id; }
+	void SetSelectedObjP(int64 id) { _selectedIdP = id; }
 
-	int32 GetSelectedIdH() { return _selectedIdH; }
-	int32 GetSelectedIdP() { return _selectedIdP; }
+	int64 GetSelectedIdH() { return _selectedIdH; }
+	int64 GetSelectedIdP() { return _selectedIdP; }
 	
 	shared_ptr<LogWindow> GetLog();
 
 private:
 	
 	 bool _hiearchyWindow = false;
-	 int32 _selectedIdH = -1;
-	 int32 _selectedIdP = -1; 
+	 int64 _selectedIdH = -1;
+	 int64 _selectedIdP = -1;
 
 private:
 	unordered_map<string, shared_ptr<class EditorWindow>> _editorWindows;
